src/backend/requirements.txt
# python 3.10.12
ucimlrepo==0.0.7
numpy==1.26.2
pandas==2.2.2
scikit-learn==1.5.1
fastapi==0.115.12
fastapi-cli==0.0.4
uvicorn==0.30.3
opencv-python==4.10.0.84
pymongo==4.10.1
gradio==5.25.2
pyyaml==6.0.2
itsdangerous==2.2.0
authlib==1.5.2
kafka-python==2.2.10
PyJWT==2.10.1

# bo sung
python-dotenv==1.1.1
httpx==0.28.1
aiokafka==0.12.0
minio==7.2.18
pyarrow==21.0.0
fastparquet==2024.11.0
scikit-optimize==0.10.2
